+++ src/storage/store/store_sync/memory_store.rs
@@ -5,7 +5,7 @@ use std::sync::Mutex;
 
 use crate::{
     array::MaybeBytes,
-    byte_range::{ByteOffset, ByteRange},
+    byte_range::{ByteOffset, ByteRange, InvalidByteRangeError},
     storage::{
         store_lock::{DefaultStoreLocks, StoreKeyMutex, StoreLocks},
         store_set_partial_values, ListableStorageTraits, ReadableStorageTraits,
@@ -101,6 +101,9 @@ impl ReadableStorageTraits for MemoryStore {
             for byte_range in byte_ranges {
                 let start = usize::try_from(byte_range.start(data.len() as u64)).unwrap();
                 let end = usize::try_from(byte_range.end(data.len() as u64)).unwrap();
+                if end > data.len() {
+                    return Err(InvalidByteRangeError.into());
+                }
                 let bytes = data[start..end].to_vec();
                 out.push(bytes);
             }
@@ -219,101 +222,11 @@ mod tests {
     use std::error::Error;
 
     #[test]
-    fn memory_set() -> Result<(), Box<dyn Error>> {
+    fn memory() -> Result<(), Box<dyn Error>> {
         let store = MemoryStore::new();
-        let key = "a/b".try_into()?;
-        store.set(&key, &[0, 1, 2])?;
-        assert_eq!(store.get(&key)?.unwrap(), &[0, 1, 2]);
-        store.set_partial_values(&[StoreKeyStartValue::new(key.clone(), 1, &[3, 4])])?;
-        assert_eq!(store.get(&key)?.unwrap(), &[0, 3, 4]);
-
-        assert_eq!(
-            store
-                .get_partial_values_key(&key, &[ByteRange::FromStart(1, None)])?
-                .unwrap()
-                .first()
-                .unwrap(),
-            &[3, 4]
-        );
-
-        assert!(store
-            .get_partial_values_key(&"a/b/c".try_into()?, &[ByteRange::FromStart(1, None)])?
-            .is_none());
-
-        assert_eq!(
-            store
-                .get_partial_values(&[StoreKeyRange::new(
-                    key.clone(),
-                    ByteRange::FromStart(1, None)
-                )])?
-                .first()
-                .unwrap()
-                .as_ref()
-                .unwrap(),
-            &[3, 4]
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn memory_list() -> Result<(), Box<dyn Error>> {
-        let store = MemoryStore::new();
-
-        store.set(&"a/b".try_into()?, &[0])?;
-        store.set(&"a/c".try_into()?, &[0, 0])?;
-        store.set(&"a/d/e".try_into()?, &[])?;
-        store.set(&"a/d/f".try_into()?, &[])?;
-        store.erase(&"a/d/e".try_into()?)?;
-        assert_eq!(
-            store.list()?,
-            &["a/b".try_into()?, "a/c".try_into()?, "a/d/f".try_into()?]
-        );
-        assert_eq!(
-            store.list_prefix(&"a/".try_into()?)?,
-            &["a/b".try_into()?, "a/c".try_into()?, "a/d/f".try_into()?]
-        );
-        assert_eq!(
-            store.list_prefix(&"a/d/".try_into()?)?,
-            &["a/d/f".try_into()?]
-        );
-        assert_eq!(
-            store.list_prefix(&"".try_into()?)?,
-            &["a/b".try_into()?, "a/c".try_into()?, "a/d/f".try_into()?]
-        );
-
-        assert_eq!(store.list_prefix(&"b/".try_into()?)?, &[]);
-
-        assert_eq!(store.size()?, 3);
-        assert_eq!(store.size_prefix(&"a/".try_into().unwrap())?, 3);
-        assert_eq!(store.size_key(&"a/b".try_into().unwrap())?, Some(1));
-        Ok(())
-    }
-
-    #[test]
-    fn memory_list_dir() -> Result<(), Box<dyn Error>> {
-        let store = MemoryStore::new();
-        store.set(&"a/b".try_into()?, &[])?;
-        store.set(&"a/c".try_into()?, &[])?;
-        store.set(&"a/d/e".try_into()?, &[])?;
-        store.set(&"a/f/g".try_into()?, &[])?;
-        store.set(&"a/f/h".try_into()?, &[])?;
-        store.set(&"b/c/d".try_into()?, &[])?;
-
-        let list_dir = store.list_dir(&StorePrefix::root())?;
-        assert_eq!(list_dir.prefixes(), &["a/".try_into()?, "b/".try_into()?,]);
-
-        let list_dir = store.list_dir(&"a/".try_into()?)?;
-
-        assert_eq!(list_dir.keys(), &["a/b".try_into()?, "a/c".try_into()?,]);
-        assert_eq!(
-            list_dir.prefixes(),
-            &["a/d/".try_into()?, "a/f/".try_into()?,]
-        );
-
-        store.erase_prefix(&"b/".try_into()?)?;
-        let list_dir = store.list_dir(&StorePrefix::root())?;
-        assert_eq!(list_dir.prefixes(), &["a/".try_into()?,]);
-
+        super::super::test_util::store_write(&store)?;
+        super::super::test_util::store_read(&store)?;
+        super::super::test_util::store_list(&store)?;
         Ok(())
     }
 }